feat: `mssql` collection enhancements and bug fixes (#265)
* address bugganizers due to sql server versions

* Update to include associated table_name to object

* Further refine listing

* re-add object count

* MSSQL-Adding warnings if perfmon file doesn't exist and create an empty file (#258)

* Added warning & create empty perfmon file if one cannot be found

* Updated after testing

* Update help and license information

* SQL Server : Adding help menu to bat files (#263)

* Added warning & create empty perfmon file if one cannot be found

* clean up test commands

* uncommented temp file cleanup line

* Add help menu to bat files

* added help menu in runAssessment bat file

---------


---------

+++ scripts/collector/sqlserver/dma_sqlserver_perfmon_dataset.ps1
@@ -28,8 +28,7 @@
 .EXAMPLE
     C:\dma_sqlserver_perfmon_dataset.ps1 -operation create -mssqlInstanceName [named Instance] / $null
 .NOTES
-    Author: Shane Borden
-    Date:   February 28, 2023    
+    https://googlecloudplatform.github.io/database-assessment/
 #>
 param (
 	[Parameter(
@@ -380,6 +379,8 @@ param(
 		logman.exe delete -n $dataSet
 	}
 }
+function Yellow{    process { Write-Host $_ -ForegroundColor Yellow }}
+
 function CollectDMAPerfmonDataSet
 {
 param(
@@ -401,6 +402,7 @@ param(
 	if (Test-Path -Path $env:SystemDrive\PerfLogs\Admin\Google-DMA-SQLServerDataSet\*$dataSet*.csv) {
 		Write-Output ""
 		Write-Output "Moving perfmon datafiles to the $env:TEMP Directory Without Header"
+		$fileExists = $true
 		foreach($file in Get-ChildItem -Path $env:SystemDrive\PerfLogs\Admin\Google-DMA-SQLServerDataSet\*$dataset*.csv)
 		{
 			$tempFileName = Split-Path $file -leaf
@@ -414,12 +416,17 @@ param(
 		}
 	} else {
 		Write-Output ""
-		Write-Output "No Perfmon Files exist in the $env:SystemDrive\PerfLogs\Admin\Google-DMA-SQLServerDataSet Directory"
+		Write-Output "No Perfmon Files exist in the $env:SystemDrive\PerfLogs\Admin\Google-DMA-SQLServerDataSet Directory. Continuing without Perform file." | Yellow
+		$fileExists = $false
 	}
 
 	Write-Output "Concatenating and adding header to perfmon files to $outputFileName" 
+	if ($fileExists)  {
 	((Get-Content -Path $PSScriptRoot\perfmon_header.csv, $env:TEMP\PKEY_*$dataSet*.csv -Raw ) -replace ',','|') | Set-Content -Encoding utf8 -NoNewline -Path $outputDir\$outputFileName
-	
+	}
+	else {
+	((Get-Content -Path $PSScriptRoot\perfmon_header.csv -Raw ) -replace ',','|') | Set-Content -Encoding utf8 -NoNewline -Path $outputDir\$outputFileName
+	}	
 	if (Test-Path -Path $outputDir\$outputFileName) {
 		Write-Output "Clean up Temp File area."
 		Remove-Item -Path $env:TEMP\*$dataSet*.csv
